+++ mcp_tools/web_app_security/enhanced_web_app_security.py
@@ -3,73 +3,6 @@
 from typing import Dict, Any, Optional
 
 def register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger, HexStrikeColors):
-    @mcp.tool()
-    def dirsearch_scan(url: str, extensions: str = "php,html,js,txt,xml,json",
-                      wordlist: str = "/usr/share/wordlists/dirsearch/common.txt",
-                      threads: int = 30, recursive: bool = False, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Dirsearch for advanced directory and file discovery with enhanced logging.
-
-        Args:
-            url: The target URL
-            extensions: File extensions to search for
-            wordlist: Wordlist file to use
-            threads: Number of threads to use
-            recursive: Enable recursive scanning
-            additional_args: Additional Dirsearch arguments
-
-        Returns:
-            Advanced directory discovery results
-        """
-        data = {
-            "url": url,
-            "extensions": extensions,
-            "wordlist": wordlist,
-            "threads": threads,
-            "recursive": recursive,
-            "additional_args": additional_args
-        }
-        logger.info(f"📁 Starting Dirsearch scan: {url}")
-        result = hexstrike_client.safe_post("api/tools/dirsearch", data)
-        if result.get("success"):
-            logger.info(f"✅ Dirsearch scan completed for {url}")
-        else:
-            logger.error(f"❌ Dirsearch scan failed for {url}")
-        return result
-
-    @mcp.tool()
-    def katana_crawl(url: str, depth: int = 3, js_crawl: bool = True,
-                    form_extraction: bool = True, output_format: str = "json",
-                    additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Katana for next-generation crawling and spidering with enhanced logging.
-
-        Args:
-            url: The target URL to crawl
-            depth: Crawling depth
-            js_crawl: Enable JavaScript crawling
-            form_extraction: Enable form extraction
-            output_format: Output format (json, txt)
-            additional_args: Additional Katana arguments
-
-        Returns:
-            Advanced web crawling results with endpoints and forms
-        """
-        data = {
-            "url": url,
-            "depth": depth,
-            "js_crawl": js_crawl,
-            "form_extraction": form_extraction,
-            "output_format": output_format,
-            "additional_args": additional_args
-        }
-        logger.info(f"⚔️  Starting Katana crawl: {url}")
-        result = hexstrike_client.safe_post("api/tools/katana", data)
-        if result.get("success"):
-            logger.info(f"✅ Katana crawl completed for {url}")
-        else:
-            logger.error(f"❌ Katana crawl failed for {url}")
-        return result
 
     @mcp.tool()
     def gau_discovery(domain: str, providers: str = "wayback,commoncrawl,otx,urlscan",
