+++ mcp_tools/web_crawl/katana.py
@@ -1,0 +1,39 @@
+# mcp_tools/web_crawl/katana.py
+
+
+from typing import Dict, Any, Optional
+
+def register_katana_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def katana_crawl(url: str, depth: int = 3, js_crawl: bool = True,
+                    form_extraction: bool = True, output_format: str = "json",
+                    additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Katana for next-generation crawling and spidering with enhanced logging.
+
+        Args:
+            url: The target URL to crawl
+            depth: Crawling depth
+            js_crawl: Enable JavaScript crawling
+            form_extraction: Enable form extraction
+            output_format: Output format (json, txt)
+            additional_args: Additional Katana arguments
+
+        Returns:
+            Advanced web crawling results with endpoints and forms
+        """
+        data = {
+            "url": url,
+            "depth": depth,
+            "js_crawl": js_crawl,
+            "form_extraction": form_extraction,
+            "output_format": output_format,
+            "additional_args": additional_args
+        }
+        logger.info(f"⚔️  Starting Katana crawl: {url}")
+        result = hexstrike_client.safe_post("api/tools/katana", data)
+        if result.get("success"):
+            logger.info(f"✅ Katana crawl completed for {url}")
+        else:
+            logger.error(f"❌ Katana crawl failed for {url}")
+        return result
